requirements-prod.txt
# Production dependencies for Local Call Center AI

# API Framework
fastapi==0.121.3
uvicorn[standard]==0.32.1
gunicorn==23.0.0
python-multipart==0.0.17

# Async Support
aiohttp==3.11.11
aiofiles==24.1.0
httpx==0.28.1

# WebSocket
websockets==13.1
python-socketio==5.11.4

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0

# Redis
redis==5.2.1
hiredis==3.0.0

# Message Queue
celery==5.4.0
kombu==5.4.2

# AI/ML Models
torch==2.5.1
transformers==4.47.1
accelerate==1.2.1
optimum==1.24.0
sentencepiece==0.2.0
protobuf==5.29.3

# Audio Processing
numpy==1.26.4
scipy==1.14.1
librosa==0.10.2
soundfile==0.12.1
pydub==0.25.1
webrtcvad==2.0.10

# Telephony
twilio==9.3.13

# Monitoring & Observability
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-aiohttp-client==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-redis==0.50b0
opentelemetry-exporter-prometheus==0.50b0
opentelemetry-exporter-otlp==1.29.0

# Logging
structlog==24.6.0
python-json-logger==3.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
cryptography==44.0.0

# Validation
pydantic==2.12.4
pydantic-settings==2.7.0
email-validator==2.2.0

# Rate Limiting
slowapi==0.1.9

# Caching
aiocache==0.12.3
msgpack==1.1.0

# Health Checks
py-healthcheck==3.0.1

# Configuration
pyyaml==6.0.2
python-dotenv==1.0.1
dynaconf==3.2.6

# Testing (for production tests)
pytest==8.3.4
pytest-asyncio==0.25.1
pytest-cov==6.0.0
pytest-benchmark==5.1.0

# Code Quality
mypy==1.14.1
black==24.10.0
flake8==7.1.1
pylint==3.3.3
bandit==1.8.0
safety==3.2.11

# Performance
py-spy==0.4.0
memory-profiler==0.61.0
line-profiler==4.2.0

# Utilities
tenacity==9.0.0
backoff==2.2.1
python-dateutil==2.9.0
pytz==2024.2
orjson==3.10.13
ujson==5.10.0

# Documentation
mkdocs==1.6.1
mkdocs-material==9.5.49
mkdocstrings[python]==0.27.0

# Deployment
supervisor==4.2.6
circus==0.18.0
